Refactored editable to get into a separate file and added uniq class names'
