+++ lib/exercises_index.js
@@ -3,57 +3,8 @@ var $ = require('jquery');
 require('../css/main.scss');
 
 var Exercises = require('./exercises');
+var Editable = require('./editable')
 
-//editable ----------------------------------------------------------------------
-
-function makeEditable(table) {
-  $('.'+table).on('click', 'td', function (event) {
-    var className    = classNameClickedRow(this);
-    var oldName      = recordCellHtmlName(className);
-    var oldCalories  = recordCellHtmlCalories(className);
-    var el = event.target;
-    addEventListenerForEnter(table, className, oldName, oldCalories, el);
-    addEventListenerForMousedown(table, className, oldName, oldCalories, el);
-  });
-};
-
-function recordCellHtmlName (className) {
-  return $('.'+className).children('td')[0].innerHTML;
-};
-
-function recordCellHtmlCalories (className) {
-  return $('.'+className).children('td')[1].innerHTML;
-};
-
-function classNameClickedRow (clicked) {
-  return clicked.parentElement.className;
-};
-
-function addEventListenerForEnter (table, className, oldName, oldCalories, el) {
-  document.getElementById(table + '-table').addEventListener('keydown', function (event) {
-    if(event.keyCode == 13 || event.keyCode == 9) {
-      event.preventDefault();
-      var newName      = recordCellHtmlName(className);
-      var newCalories  = recordCellHtmlCalories(className);
-      editExercisesStorage(newName, newCalories, oldName, oldCalories);
-      el.blur();
-    }
-    }, true);
-};
-
-function addEventListenerForMousedown (table, className, oldName, oldCalories, el) {
-  document.addEventListener('mousedown', function (event) {
-    var newName      = recordCellHtmlName(className);
-    var newCalories  = recordCellHtmlCalories(className);
-    editExercisesStorage(newName, newCalories, oldName, oldCalories);
-    el.blur();
-  }, true);
-}
-
-function editExercisesStorage (newName, newCalories, oldName, oldCalories) {
-  var exercises  = new Exercises(newName, newCalories);
-  exercises.edit(oldName, oldCalories);
-};
 
 TableInitialize("exercises");
 
@@ -73,7 +24,7 @@ function TableInitialize (table) {
       }
     };
     $('tr').each(function(i) {
-      $(this).removeClass('new-row');
+      $(this).removeAttr('class');
       $( this ).addClass( "row" + i );
     });
   };
@@ -81,9 +32,11 @@ function TableInitialize (table) {
 
 
 function addToTable(name, calories, table) {
-  var row = '<tr id="editable" class="new-row" ><td contenteditable="true">' +name+ '</td><td contenteditable="true">' +calories+ '</td><td><a href="#" class="delete-'+table+' material-icons">delete</a></td></tr>'
+  var tempUniqClassName = (name+calories+table).replace(/\s+/g, '')
+  var row = '<tr id="editable" class="'+tempUniqClassName+'" ><td contenteditable="true">' +name+ '</td><td contenteditable="true">' +calories+ '</td><td><a href="#" class="delete-'+table+' material-icons">delete</a></td></tr>'
   $('.'+table+' > tbody').prepend(row);
-  makeEditable(table);
+  var editable  = new Editable();
+  editable.make(table);
 };
 
 function deleteRow (row, table) {
