+++ lib/foods_index.js
@@ -4,57 +4,7 @@ var _ = require('lodash');
 require('../css/main.scss');
 
 var Food = require('./foods');
-
-//editable ----------------------------------------------------------------------
-
-function makeEditable(table) {
-  $('.'+table).on('click', 'td', function (event) {
-    var className    = classNameClickedRow(this);
-    var oldName      = recordCellHtmlName(className);
-    var oldCalories  = recordCellHtmlCalories(className);
-    var el = event.target;
-    addEventListenerForEnter(table, className, oldName, oldCalories, el);
-    addEventListenerForMousedown(table, className, oldName, oldCalories, el);
-  });
-};
-
-function recordCellHtmlName (className) {
-  return $('.'+className).children('td')[0].innerHTML;
-};
-
-function recordCellHtmlCalories (className) {
-  return $('.'+className).children('td')[1].innerHTML;
-};
-
-function classNameClickedRow (clicked) {
-  return clicked.parentElement.className;
-};
-
-function addEventListenerForEnter (table, className, oldName, oldCalories, el) {
-  document.getElementById(table + '-table').addEventListener('keydown', function (event) {
-    if(event.keyCode == 13 || event.keyCode == 9) {
-      event.preventDefault();
-      var newName      = recordCellHtmlName(className);
-      var newCalories  = recordCellHtmlCalories(className);
-      editFoodStorage(newName, newCalories, oldName, oldCalories);
-      el.blur();
-    }
-    }, true);
-};
-
-function addEventListenerForMousedown (table, className, oldName, oldCalories, el) {
-  document.addEventListener('mousedown', function (event) {
-    var newName      = recordCellHtmlName(className);
-    var newCalories  = recordCellHtmlCalories(className);
-    editFoodStorage(newName, newCalories, oldName, oldCalories);
-    el.blur();
-  }, true);
-}
-
-function editFoodStorage (newName, newCalories, oldName, oldCalories) {
-  var food  = new Food(newName, newCalories);
-  food.edit(oldName, oldCalories);
-};
+var Editable = require('./editable')
 
 TableInitialize("foods");
 
@@ -74,7 +24,7 @@ function TableInitialize (table) {
       }
     };
     $('tr').each(function(i) {
-      $(this).removeClass('new-row');
+      $(this).removeAttr('class');
       $( this ).addClass( "row" + i );
     });
   };
@@ -82,9 +32,11 @@ function TableInitialize (table) {
 
 
 function addToTable(name, calories, table) {
-  var row = '<tr id="editable" class="new-row" ><td contenteditable="true">' +name+ '</td><td contenteditable="true">' +calories+ '</td><td><a href="#" class="delete-'+table+' material-icons">delete</a></td></tr>'
+  var tempUniqClassName = (name+calories+table).replace(/\s+/g, '');
+  var row = '<tr id="editable" class="'+tempUniqClassName+'" ><td contenteditable="true">' +name+ '</td><td contenteditable="true">' +calories+ '</td><td><a href="#" class="delete-'+table+' material-icons">delete</a></td></tr>'
   $('.'+table+' > tbody').prepend(row);
-  makeEditable(table);
+  var editable  = new Editable();
+  editable.make(table);
 };
 
 function deleteRow (row, table) {
