+++ lib/index.js
@@ -11,117 +11,10 @@ var Lunch = require('./lunch');
 var Dinner = require('./dinner');
 var Snacks = require('./snacks');
 var Exercise = require('./exercise')
+var Editable = require('./editable')
 
-//editable ----------------------------------------------------------------------
 
 initializeDays();
-
-function makeEditable(table) {
-  $('.'+table).on('click', 'td', function (event) {
-    var className    = classNameClickedRow(this);
-    var oldName      = recordCellHtmlName(className);
-    var oldCalories  = recordCellHtmlCalories(className);
-    var el = event.target;
-    addEventListenerForEnter(table, className, oldName, oldCalories, el);
-    addEventListenerForMousedown(table, className, oldName, oldCalories, el);
-  });
-};
-
-function recordCellHtmlName (className) {
-  return $('.'+className).children('td')[0].innerHTML;
-};
-
-function recordCellHtmlCalories (className) {
-  return $('.'+className).children('td')[1].innerHTML;
-};
-
-function classNameClickedRow (clicked) {
-  return $(clicked).parents('tr').attr('class');
-};
-
-function addEventListenerForEnter (table, className, oldName, oldCalories, el) {
-  document.getElementById(table + '-table').addEventListener('keydown', function (event) {
-    if(event.keyCode == 13 || event.keyCode == 9) {
-      event.preventDefault();
-      var newName      = recordCellHtmlName(className);
-      var newCalories  = recordCellHtmlCalories(className);
-      editCorrectTable(table, newName, newCalories, oldName, oldCalories);
-      el.blur();
-    }
-  }, true);
-};
-
-function addEventListenerForMousedown (table, className, oldName, oldCalories, el) {
-  document.addEventListener('mousedown', function (event) {
-    var newName      = recordCellHtmlName(className);
-    var newCalories  = recordCellHtmlCalories(className);
-    editCorrectTable(table, newName, newCalories, oldName, oldCalories);
-    el.blur();
-  }, true);
-}
-
-function editCorrectTable (table, newName, newCalories, oldName, oldCalories) {
-    switch(table) {
-      case 'diary-foods':
-        editFoodStorage(newName, newCalories, oldName, oldCalories);
-        break;
-      case 'diary-exercises':
-        editExercisesStorage(newName, newCalories, oldName, oldCalories);
-        break;
-      case 'breakfast':
-        editBreakfastStorage(newName, newCalories, oldName, oldCalories);
-        break;
-      case 'lunch':
-        editLunchStorage(newName, newCalories, oldName, oldCalories);
-        break;
-      case 'dinner':
-        editDinnerStorage(newName, newCalories, oldName, oldCalories);
-        break;
-      case 'snacks':
-        editSnacksStorage(newName, newCalories, oldName, oldCalories);
-        break;
-      case 'exercise':
-        editExerciseStorage(newName, newCalories, oldName, oldCalories);
-        break;
-      }
-}
-
-function editFoodStorage (newName, newCalories, oldName, oldCalories) {
-  var food  = new Food(newName, newCalories);
-  food.edit(oldName, oldCalories);
-};
-
-function editExercisesStorage (newName, newCalories, oldName, oldCalories) {
-  var exercises  = new Exercises(newName, newCalories);
-  exercises.edit(oldName, oldCalories);
-};
-
-function editBreakfastStorage (newName, newCalories, oldName, oldCalories) {
-  var breakfast  = new Breakfast(newName, newCalories);
-  breakfast.edit($('#date').text(), oldName, oldCalories);
-};
-
-function editLunchStorage (newName, newCalories, oldName, oldCalories) {
-  var lunch  = new Lunch(newName, newCalories);
-  lunch.edit($('#date').text(), oldName, oldCalories);
-};
-
-function editDinnerStorage (newName, newCalories, oldName, oldCalories) {
-  var dinner  = new Dinner(newName, newCalories);
-  dinner.edit($('#date').text(), oldName, oldCalories);
-};
-
-function editSnacksStorage (newName, newCalories, oldName, oldCalories) {
-  var snacks  = new Snacks(newName, newCalories);
-  snacks.edit($('#date').text(), oldName, oldCalories);
-};
-
-function editExerciseStorage (newName, newCalories, oldName, oldCalories) {
-  var exercise  = new Exercise(newName, newCalories);
-  exercise.edit($('#date').text(), oldName, oldCalories);
-};
-
-
 makeTables();
 
 // food to meal table
@@ -159,15 +52,27 @@ function TableInitialize (table) {
 };
 
 function addToTable(name, calories, table) {
-  var row = '<tr id="editable" class="new-row" ><td contenteditable="true">' +name+ '</td><td contenteditable="true">' +calories+ '</td><td><a href="#" class="delete-'+table+' material-icons">delete</a></td></tr>'
+  var tempUniqClassName = (name+calories+table).replace(/\s+/g, '')
+  var row = '<tr id="editable" class="'+tempUniqClassName+'" ><td contenteditable="true">' +name+ '</td><td contenteditable="true">' +calories+ '</td><td><a href="#" class="delete-'+table+' material-icons">delete</a></td></tr>'
   $('.'+table+' > tbody').prepend(row);
-  makeEditable(table);
+  var editable  = new Editable();
+  editable.make(table);
   calculateCaloriesTotal(table);
   calculateRemainingCalories(table);
   caloriesConsumedTotal();
   caloriesBurnedTotal();
   caloriesRemainingTotal();
 };
+
+function addToCheckBoxTable(name, calories, table) {
+  var row = '<tr id="editable" class="new-row" ><td contenteditable="true">' +name+ '</td><td contenteditable="true">' +calories+ '</td><td><input type="checkbox" id="'+name+'"/><label for="'+name+'"></label></td></tr>'
+  $('.'+table+' > tbody').prepend(row);
+  var editable  = new Editable();
+  editable.make(table);
+};
+
+
+// Calculations ------------------------------
 
 function calculateCaloriesTotal(table) {
     var sum = addCalories(table);
@@ -228,6 +133,7 @@ function addCalories(table) {
 }
 
 
+// Total Calories table _________________________
 
 function addTotalCaloriesConsumed() {
   return addCalories('breakfast') + addCalories('lunch') + addCalories('dinner') + addCalories('snacks')
@@ -264,11 +170,7 @@ function caloriesRemainingTotal() {
   }
 }
 
-function addToCheckBoxTable(name, calories, table) {
-  var row = '<tr id="editable" class="new-row" ><td contenteditable="true">' +name+ '</td><td contenteditable="true">' +calories+ '</td><td><input type="checkbox" id="'+name+'"/><label for="'+name+'"></label></td></tr>'
-  $('.'+table+' > tbody').prepend(row);
-  makeEditable(table);
-};
+// Delete-------------------------------------
 
 function deleteRow (row, table) {
   var doomedRow = $(row).parents('tr')
@@ -328,6 +230,8 @@ function deleteLocalStorage(name, calories, table) {
     }
 }
 
+//  Storing checked box info ------------------------
+
 function checkedBoxInfo (table, fromTable) {
 var meals = $('.diary-'+fromTable+' input[type=checkbox]:checked');
   _.forEach(meals, function(meal, index){
@@ -362,7 +266,6 @@ function storeCheckedBoxInfo (table) {
     break;
   }
 }
-
 // Breakfast Diary-------------------------------------------------------------------------
 
 $('.breakfast').on('click', '.delete-breakfast', function (event) {
@@ -431,7 +334,8 @@ $('.add-exercise').on('click', function (e) {
 })
 
 
-// Filter
+// Filter----------------------------------------
+
 function displayDate(date) {
   $('#date').text(date);
 }
